+++ modules/volumes/main.tf
@@ -30,7 +30,7 @@ resource "oci_core_volume" "volume" {
   dynamic "block_volume_replicas" {
     for_each = each.value.disable_replicas ? {} : each.value.cross_ad_replicas
     content {
-      availability_domain = block_volume_replicas.value.replica_region
+      availability_domain = block_volume_replicas.value.destination_availability_domain
       display_name        = block_volume_replicas.value.replica_name
     }
   }